Moved resource cleanup manager to its own file
